Move GUI bootstraping to gui/BootstrapApplication.
This removes some more Qt::Widget dependencies from core.
