+++ src/CMakeLists.txt
@@ -100,6 +100,7 @@ set(keepassx_SOURCES
         gui/styles/light/LightStyle.cpp
         gui/AboutDialog.cpp
         gui/Application.cpp
+        gui/BootstrapApplication.cpp
         gui/CategoryListWidget.cpp
         gui/Clipboard.cpp
         gui/CloneDialog.cpp
